requirements.txt
fastapi
uvicorn
strands-agents
strands-agents-tools
strands-agents-builder
mcp
httpx